+++ CHANGELOG.md
@@ -11,7 +11,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - ![STAT] Correct the refreshing of models with Julian Easter [#465](https://github.com/jdemetra/jdplus-main/issues/465)
 - ![STAT] Correct a bug in the estimation of an ARIMA model using multiple starting points (X13 only, under very specific conditions)
+- ![STAT] Add a missing revision policy in the cruncher [#495](https://github.com/jdemetra/jdplus-main/issues/495)
 - ![IO] Use the same name for "SAProcessing" in java and R
+- ![IO] Add missing information in the exports of the residuals of REGARIMA models 
 - ![UI] Correct the "Check Last" user interface [#511](https://github.com/jdemetra/jdplus-main/issues/511)
 - ![UI] Correct the "SI-Ratio" panel (last values not shown)  [#518](https://github.com/jdemetra/jdplus-main/issues/518)